+++ packages/methods/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the library will be documented in this file.
 ## vX.X.X (Month DD, YYYY)
 
 - Change `@formisch/core` to vX.X.X
+- Fix bug with `elements` state by resetting it to `initialElements` state
 
 ## v0.2.0 (July 27, 2025)